op mizantrop

+++ tables/views/client.py
@@ -17,45 +17,56 @@ def client_create(request, company_pk):
     company = get_object_or_404(Company, pk=company_pk)
     form = ClientForm(request.POST or None);
     user = request.user
-    clients = Client.objects.filter(company=company)
-    if request.method == 'POST':
-        if form.is_valid():
-            client = form.save(commit=False)
-            client.company = company
-            client.save()
-            messages.success(request, 'Клиент успешно добавлен')
-            return redirect(reverse('tables:client-create', kwargs={'company_pk': company.pk}))
-    
-    return render(request, 'tables/client/client_create.html', {'company': company, 'form': form, 'clients': clients})
+
+    if user.profile.in_company(company.pk) and (user.profile.is_manager() or user.profile.is_agent()):
+        if request.method == 'POST':
+            if form.is_valid():
+                client = form.save(commit=False)
+                client.company = company
+                client.save()
+                messages.success(request, 'Клиент успешно добавлен')
+                return redirect(reverse('tables:client-create', kwargs={'company_pk': company.pk}))
+        
+        return render(request, 'tables/client/client_create.html', {'company': company, 'form': form})
+    else:
+        return redirect(reverse('tables:index'))
 
 
 @login_required
 def clients(request, company_pk):
     company = get_object_or_404(Company, pk=company_pk)
     user = request.user
-    return render(request, 'tables/client/clients.html', {'company': company})
+
+    if user.profile.in_company(company.pk) and (user.profile.is_manager() or user.profile.is_agent()):
+        return render(request, 'tables/client/clients.html', {'company': company})
+    else:
+        return redirect(reverse('tables:index'))
 
 
 @login_required
 def clients_list(request, company_pk):
     user = request.user
     company = get_object_or_404(Company, pk=company_pk)
-    clients = Client.objects.filter(company=company)
-    result = []
 
-    for client in clients:
-        result.append({
-            'pk': client.pk,
-            'name': client.name,
-            'description': client.description,
-            'address': client.address,
-            'telephone_number': client.telephone_number,
-            'actions': {
-                'edit': reverse('tables:client-edit', kwargs={'company_pk': company.pk, 'client_pk':client.pk}),
-                'delete': reverse('tables:client-delete', kwargs={'company_pk': company.pk, 'client_pk':client.pk})
-            }
-        })
-    return result
+    if user.profile.in_company(company.pk) and (user.profile.is_manager() or user.profile.is_agent()):
+        clients = Client.objects.filter(company=company)
+        result = []
+
+        for client in clients:
+            result.append({
+                'pk': client.pk,
+                'name': client.name,
+                'description': client.description,
+                'address': client.address,
+                'telephone_number': client.telephone_number,
+                'actions': {
+                    'edit': reverse('tables:client-edit', kwargs={'company_pk': company.pk, 'client_pk':client.pk}),
+                    'delete': reverse('tables:client-delete', kwargs={'company_pk': company.pk, 'client_pk':client.pk})
+                }
+            })
+        return result
+    else:
+        return redirect(reverse('tables:index'))
 
 
 
@@ -68,42 +79,45 @@ def clients_json(request, company_pk):
 @login_required
 def client_edit(request, company_pk, client_pk):
     company = get_object_or_404(Company, pk=company_pk)
-    client = get_object_or_404(Client, pk=client_pk)
+    client = get_object_or_404(Client, company=company, pk=client_pk)
     form = ClientForm(request.POST or None, instance=client);
     user = request.user
-    clients = Client.objects.filter(company=company)
-    if request.method == 'POST':
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Изменения сохранены')
-            return redirect(reverse('tables:client-edit', kwargs={'company_pk': company.pk, 'client_pk':client.pk }))
-    context = {
-        'company': company,
-        'client': client,
-        'form': form
-    }
-    return render(request, 'tables/client/client_edit.html', context)
+
+    if user.profile.in_company(company.pk) and (user.profile.is_manager() or user.profile.is_agent()):
+        if request.method == 'POST':
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Изменения сохранены')
+                return redirect(reverse('tables:client-edit', kwargs={'company_pk': company.pk, 'client_pk':client.pk }))
+        context = {
+            'company': company,
+            'client': client,
+            'form': form
+        }
+        return render(request, 'tables/client/client_edit.html', context)
+    else:
+        return redirect(reverse('tables:index'))
 
 
 @login_required
 def client_delete(request, company_pk, client_pk):
     company = get_object_or_404(Company, pk=company_pk)
-    client = get_object_or_404(Client, pk=client_pk)
+    client = get_object_or_404(Client, company=company, pk=client_pk)
     user = request.user
     data = dict()
 
-    if request.method == 'POST':
-        client.delete()
-        data['form_is_valid'] = True
-        data['clients_list'] = clients_list(request, company_pk)
-    else:
-        context = {
-            'client': client,
-            'company': company,
-        }
-        data['html_form'] = render_to_string('tables/client/client_delete_form.html', context, request=request)
-    return JsonResponse(data)
-
+    if user.profile.in_company(company.pk) and (user.profile.is_manager() or user.profile.is_agent()):
+        if request.method == 'POST':
+            client.delete()
+            data['form_is_valid'] = True
+        else:
+            context = {
+                'client': client,
+                'company': company,
+            }
+            data['html_form'] = render_to_string('tables/client/client_delete_form.html', context, request=request)
+        return JsonResponse(data)
+    return redirect(reverse('tables:index'))
 
 
 @login_required
@@ -112,16 +126,18 @@ def clients_delete(request, company_pk):
     user = request.user
     data = dict()
 
-    if request.method == 'POST':
-        ids = request.POST.get('ids').split(',')
-        cleaned_ids = []
-        for _id in ids:
-            try:
-                number = int(_id)
-                cleaned_ids.append(number)
-            except:
-                return HttpResponseForbidden()
-        Client.objects.filter(pk__in=cleaned_ids, company=company).delete()
-        data['form_is_valid'] = True
-        data['clients_list'] = clients_list(request, company_pk)
-    return JsonResponse(data)
+    if user.profile.in_company(company.pk) and (user.profile.is_manager() or user.profile.is_agent()):
+        if request.method == 'POST':
+            ids = request.POST.get('ids').split(',')
+            cleaned_ids = []
+            for _id in ids:
+                try:
+                    number = int(_id)
+                    cleaned_ids.append(number)
+                except:
+                    return HttpResponseForbidden()
+            Client.objects.filter(pk__in=cleaned_ids, company=company).delete()
+            data['form_is_valid'] = True
+        return JsonResponse(data)
+    else:
+        return redirect(reverse('tables:index'))
